Deploy test

+++ Jumpeno.Shared/Game/Models/Point.cs
@@ -1,6 +1,6 @@
 namespace Jumpeno.Shared.Models;
 
 public record Point(
-    float X,
-    float Y
+    int X,
+    int Y
 );
